Remove SWIG support
